streamlit_app/requirements.txt
pandas
streamlit
streamlit-aggrid
streamlit-scrollable-textbox
seaborn
medcat
python-dotenv